backend/requirements.txt
python-dotenv==1.0.1
google-generativeai==0.8.3
requests==2.32.3
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-PyMongo==3.0.1
google==3.0.0
beautifulsoup4==4.13.3